migrate ReqToTokenPool back to int32

+++ python/sglang/srt/managers/schedule_batch.py
@@ -1342,7 +1342,9 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
                 self.seq_lens, last_loc
             )
 
-        self.req_to_token_pool.write((self.req_pool_indices, locs), self.out_cache_loc)
+        self.req_to_token_pool.write(
+            (self.req_pool_indices, locs), self.out_cache_loc.to(torch.int32)
+        )
 
     def filter_batch(
         self,
